Create data validation middleware
