+++ src/controllers/words.js
@@ -17,8 +17,6 @@ import {
 import { searchIgboRegexQuery, searchIgboQuery, searchEnglishRegexQuery } from './utils/queries';
 import { findWordsWithMatch } from './utils/buildDocs';
 import { createExample, executeMergeExample } from './examples';
-import { findGenericWordById } from './genericWords';
-import { findWordSuggestionById } from './wordSuggestions';
 import { sendMergedEmail } from './email';
 import { DICTIONARY_APP_URL } from '../config';
 import { findUser } from './users';
@@ -206,43 +204,9 @@ const handleSendingMergedEmail = async (result) => {
 /* Merges the existing WordSuggestion of GenericWord into either a brand
  * new Word document or merges into an existing Word document */
 export const mergeWord = async (req, res) => {
-  const { body: data } = req;
-  const { user } = req;
-  const suggestionDoc = (await findWordSuggestionById(data.id)) || (await findGenericWordById(data.id));
-
-  if (!user || (user && !user.uid)) {
-    res.status(400);
-    return res.send({ error: 'User uid is required' });
-  }
-
-  if (!suggestionDoc) {
-    res.status(400);
-    return res.send({
-      error: 'There is no associated generic word or word suggestion, double check your provided data',
-    });
-  }
-
-  if (!suggestionDoc.word) {
-    res.status(400);
-    return res.send({ error: 'The word property is missing, double check your provided data' });
-  }
-
-  if (!suggestionDoc.wordClass) {
-    res.status(400);
-    return res.send({ error: 'The word class property is missing, double check your provided data' });
-  }
-
-  if (!suggestionDoc.definitions) {
-    res.status(400);
-    return res.send({ error: 'The definition property is missing, double check your provided data' });
-  }
-
-  if (!suggestionDoc.id) {
-    res.status(400);
-    return res.send({ error: 'The id property is missing, double check your provided data' });
-  }
-
   try {
+    const { user, suggestionDoc } = req;
+
     const result = suggestionDoc.originalWordId
       ? await mergeIntoWord(suggestionDoc, user.uid)
       : await createWordFromSuggestion(suggestionDoc, user.uid);
