+++ src/routers/router.js
@@ -5,6 +5,8 @@ import { postWordSuggestion } from '../controllers/wordSuggestions';
 import { postExampleSuggestion } from '../controllers/exampleSuggestions';
 import validId from '../middleware/validId';
 import authorization from '../middleware/authorization';
+import validateWordBody from '../middleware/validateWordBody';
+import validateExampleBody from '../middleware/validateExampleBody';
 
 const router = express.Router();
 
@@ -13,7 +15,7 @@ router.get('/words/:id', validId, getWord);
 router.get('/examples', getExamples);
 router.get('/examples/:id', validId, getExample);
 
-router.post('/wordSuggestions', authorization([]), postWordSuggestion);
-router.post('/exampleSuggestions', authorization([]), postExampleSuggestion);
+router.post('/wordSuggestions', authorization([]), validateWordBody, postWordSuggestion);
+router.post('/exampleSuggestions', authorization([]), validateExampleBody, postExampleSuggestion);
 
 export default router;
